refactor: exception 패키지를 도메인 패키지 하위로 이동
