+++ src/main/java/ladder/controller/LadderGameController.java
@@ -4,7 +4,7 @@ import ladder.domain.game.LadderGame;
 import ladder.domain.rule.RandomStoolGenerator;
 import ladder.domain.ladder.Ladder;
 import ladder.domain.player.Players;
-import ladder.exception.CustomException;
+import ladder.domain.exception.CustomException;
 import ladder.view.InputView;
 import ladder.view.OutputView;
