+++ src/main/java/ladder/domain/exception/CustomException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class CustomException extends RuntimeException {
