+++ src/main/java/ladder/domain/exception/LadderLengthException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class LadderLengthException extends CustomException {
