+++ src/main/java/ladder/domain/exception/NoSuchPlayerException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class NoSuchPlayerException extends CustomException {
