+++ src/main/java/ladder/domain/exception/NotIntegerException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class NotIntegerException extends CustomException {
