+++ src/main/java/ladder/domain/exception/PlayerNameLengthException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class PlayerNameLengthException extends CustomException {
