+++ src/main/java/ladder/domain/exception/PlayerNumberException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class PlayerNumberException extends CustomException {
